+++ database/migrations/2023_05_14_165714_create_sale_details_table.php
@@ -9,9 +9,9 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('sale_details', function (Blueprint $table) {
-            $table->integer('sale_id')->nullable();
+            $table->unsignedBigInteger('sale_id')->nullable();
             $table->foreign('sale_id')->references('id')->on('sales');
-            $table->integer('product_id')->nullable();
+            $table->unsignedBigInteger('product_id')->nullable();
             $table->foreign('product_id')->references('id')->on('products');
             $table->primary(['sale_id','product_id']);
             $table->integer('quantity');
